Add Int64 and Int16 #to_param

+++ src/charms/int16_extensions.cr
@@ -1,6 +1,6 @@
 require "../lucky/allowed_in_tags"
 
-struct Int64
+struct Int16
   include ::Lucky::AllowedInTags
 
   def to_param : String
